requirements.txt
absl-py==2.4.0
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
array_record==0.8.1
asttokens==3.0.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
bddl==1.0.1
beautifulsoup4==4.14.3
bitsandbytes==0.42.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.2
cryptography==46.0.4
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
datasets==4.5.0
decorator==5.2.1
dill==0.4.0
dlimp @ git+https://github.com/moojink/dlimp_openvla@040105d256bd28866cc6620621a3d5f7b6b91b46
dm-tree==0.1.9
draccus==0.8.0
easydict==1.13
egl_probe==1.0.2
einops==0.8.2
etils==1.13.0
exceptiongroup==1.3.1
executing==2.2.1
Farama-Notifications==0.0.4
fastjsonschema==2.21.2
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
future==1.0.0
gast==0.7.0
gdown==5.2.1
gitdb==4.0.12
GitPython==3.1.46
glfw==2.10.0
google-auth==2.48.0
google-auth-oauthlib==1.2.4
google-pasta==0.2.0
grpcio==1.78.0
gymnasium==0.29.1
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
hf_egl_probe==1.0.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
hydra-core==1.3.2
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
iniconfig==2.3.0
ipython==8.38.0
jedi==0.19.2
Jinja2==3.1.6
json-numpy==2.1.1
jsonlines==4.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
jupytext==1.19.1
keras==2.15.0
kiwisolver==1.4.9
libclang==18.1.1
libero==0.1.1
llvmlite==0.46.0
mani_skill2_real2sim @ git+https://github.com/simpler-env/ManiSkill2_real2sim.git@ef7a4d4fdf4b69f2c2154db5b15b9ac8dfe10682
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
mediapy==1.2.6
mergedeep==1.3.4
ml-dtypes==0.2.0
mpmath==1.3.0
mujoco==3.5.0
multidict==6.7.1
multiprocess==0.70.18
mypy_extensions==1.1.0
nbformat==5.10.4
networkx==3.4.2
ninja==1.13.0
numba==0.63.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
omegaconf==2.3.0
opencv-python==4.13.0.92
OpenEXR==3.4.4
-e git+https://github.com/openvla/openvla.git@c8f03f48af692657d3060c19588038c7220e9af9#egg=openvla&subdirectory=../../openvla
opt_einsum==3.4.0
packaging @ file:///home/task_176104877067765/conda-bld/packaging_1761049113113/work
pandas==2.3.3
parso==0.8.5
peft==0.11.1
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
promise==2.3
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==4.25.8
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyOpenGL==3.1.10
pyparsing==3.3.2
PySocks==1.7.1
pytest==9.0.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyyaml-include==1.4.1
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.2
robomimic==0.2.0
robosuite==1.4.0
rpds-py==0.30.0
rsa==4.9.1
rtree==1.4.1
ruckig==0.15.3
safetensors==0.7.0
sapien==2.2.2
scipy==1.15.3
sentencepiece==0.1.99
sentry-sdk==2.52.0
-e git+https://github.com/simpler-env/SimplerEnv.git@06accaca93535902d408da4855f21cece12bceb7#egg=simpler_env
six==1.17.0
smmap==5.0.2
soupsieve==2.8.3
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
tensorflow==2.15.0
tensorflow-addons==0.23.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.15.0
tensorflow-graphics==2021.12.3
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.17.3
termcolor==3.3.0
thop==0.1.1.post2209072238
timm==0.9.10
tokenizers==0.19.1
toml==0.10.2
tomli==2.4.0
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tqdm==4.67.3
traitlets==5.14.3
transformers==4.40.1
transforms3d==0.4.2
trimesh==4.11.1
triton==2.2.0
typeguard==2.13.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wandb==0.24.2
wcwidth==0.6.0
Werkzeug==3.1.5
wrapt==1.14.2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
